api/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pymysql==1.0.2
cryptography==41.0.3
pydantic[email]==2.2.1
python-dotenv